a Test, Hope this works >.>
Having a little bit trouble with the GUI, but should  be done by the end
of the weekend.

The TimeLine will only view the users own Tweets, but that for testing.

once you authenticate the first time it will not ask you to do it again
going to add a sign-out button. but for now if you need to change users
you could delete the "account" file that will be in your build file.
*** need to add a checking method for users who revoked access to their
accounts manually from twitter.

** need to add some panes to the main GUI.
** Divide the GUI and function to different classes each handling a
group of related functions.